requirements.txt
fastapi 
uvicorn
sqlalchemy
google-generativeai
python-dotenv
langchain-google-genai 
langchain-core
pandas